Adjusted weapon sway when ADS and added weapon summon
Summon weapon by pressing E when looking at it. Name of the weapon also appears when the player hovers over it. The hitbox of the weapon hover is small at the moment but it is limited to only work from a distance of 30 units. The weapon also sometimes just freezes when reaching the player and it's never actually picked up...